Benchmarking argues for use of RGB over banding. Stripping out alpha only provides around 15% space savings rather than 25%. It is all zeros after all.
Banding is essentially the same as RGB, but takes slightly longer and is
slightly less compressible. There is no reason to use it unless the
introduction of filtering or run-length encoding change that.

All three are efficient enough to be usably small.

type bytes/duration -> MB per minute
-----------------------------------------
argb   469413/30340 -> 0.8853008343188009
banded 397927/30891 -> 0.7370938325815088
rgb    397828/30422 -> 0.7482710129786319

Signed-off-by: Curt Cox <curtcox@gmail.com>